+++ wltp/test/experiment_DownscaleTests.py
@@ -49,7 +49,7 @@ class ExperimentDownscale(unittest.TestCase):
         A       = np.diff(cycle)
         A       = np.append(A, 0) # Restore element lost by diff().
         A       = A / 3.6
-        P_REQ   = calcPower_required(cycle, A, mass, f0, f1, f2, f_inertial)
+        P_REQ   = calcPower_required(cycle, A, None, mass, f0, f1, f2, f_inertial)
 
         f_downscale_threshold = 0.01
         dsc_data            = class_data['downscale']