+++ wltp/test/instances_test.py
@@ -112,7 +112,7 @@ class Test(unittest.TestCase):
 
     def testModelInstance_defaultLoadCurve(self):
         json_txt = self.goodVehicle_jsonTxt % \
-            (', "full_load_curve":%s' % (json.dumps(model.default_load_curve())))
+            (', "full_load_curve":%s' % (model.json_dumps(model.default_load_curve())))
         mdl = json.loads(json_txt)
         validator = model.model_validator()